scenarios: Modularize integration tests
